Change 'ReceivedOn's type to 'int'
When 'int64' is used, its value is converted to 'Date ' format with 'UTC'
In order to its value as a epoch, change its type to 'int'

+++ buildanalysis/oebuildjobs/oebuildjobs.go
@@ -93,7 +93,7 @@ type GerritChangeInfo struct {
 	Url          string   `xml:"change>url"`
 	Changeid     string   `xml:"change>id"`
 	Patchset     PatchSet `xml:"patchSet"`
-	ReceivedOn   int64    `xml:"receivedOn"`
+	ReceivedOn   int      `xml:"receivedOn"`
 }
 
 type VerifyBuild struct {
